move dployments stuff to its separate files
