+++ roboflow/adapters/rfapi.py
@@ -6,7 +6,7 @@ from typing import Optional
 import requests
 from requests_toolbelt.multipart.encoder import MultipartEncoder
 
-from roboflow.config import API_URL, DEDICATED_DEPLOYMENT_URL, DEFAULT_BATCH_NAME, DEFAULT_JOB_NAME
+from roboflow.config import API_URL, DEFAULT_BATCH_NAME, DEFAULT_JOB_NAME
 from roboflow.util import image_utils
 
 
@@ -205,48 +205,3 @@ def _save_annotation_error(image_id, response):
     else:
         errmsg += f"bad response: {response.status_code}: {responsejson}"
     return UploadError(errmsg)
-
-
-def add_deployment(api_key, security_level, machine_type, deployment_name, inference_version):
-    url = f"{DEDICATED_DEPLOYMENT_URL}/add"
-    response = requests.post(
-        url,
-        json={
-            "api_key": api_key,
-            "security_level": security_level,
-            "machine_type": machine_type,
-            "deployment_name": deployment_name,
-            "inference_version": inference_version,
-        },
-    )
-    if response.status_code != 200:
-        raise RoboflowError(response.text)
-    result = response.json()
-    return result
-
-
-def get_deployment(api_key, deployment_id):
-    url = f"{DEDICATED_DEPLOYMENT_URL}/get"
-    response = requests.get(url, json={"api_key": api_key, "deployment_id": deployment_id})
-    if response.status_code != 200:
-        raise RoboflowError(response.text)
-    result = response.json()
-    return result
-
-
-def list_deployment(api_key):
-    url = f"{DEDICATED_DEPLOYMENT_URL}/list"
-    response = requests.get(url, json={"api_key": api_key})
-    if response.status_code != 200:
-        raise RoboflowError(response.text)
-    result = response.json()
-    return result
-
-
-def delete_deployment(api_key, deployment_id):
-    url = f"{DEDICATED_DEPLOYMENT_URL}/delete"
-    response = requests.post(url, json={"api_key": api_key, "deployment_id": deployment_id})
-    if response.status_code != 200:
-        raise RoboflowError(response.text)
-    result = response.json()
-    return result
